Ensure ifupdown2 runtime directory always exists
The /run/network directory is used by ifupdown2 for locks and state files.
This directory is created at service startup (networking). But if we want, for
example, to use ifquery before network start, it fails.
We can ensure that /run/network directory is always present with systemd-tmpfiles.

We could do it with systemd RuntimeDirectory in the networking service, but it
does not seem to be the right method with the standalone mode.

Like /run/network, we can move /etc/network/run symlink creation in ifupdown2 tmpfiles