ifreload: new --diff cli argument: only reload delta between /e/n/i
ifreload checks the delta between the user config (/e/n/i) and the
config state of the device. It's spends of lot of time checking if
the running state is valid or not.

Introducing --diff, a cli argument to only reload an interface if
the /e/n/i config was changed.

Signed-off-by: Julien Fortin <jfortin@nvidia.com>