addons: vxlan: cleaning up duplicated fdb vnifilter code
This used to be handled by ifupdown2 but the logic was moved to the kernel
thus removing the function call. For some reason the function was duplicated.

We are cleaning this up but keep the original code and a commented out func
call just in case we need it in the future.

Signed-off-by: Julien Fortin <jfortin@nvidia.com>