diff-mode: reload all clagd interface if clagd change is detected
The clagd module internally caches clagd-attribute/values and only
apply/reload the config if all interfaces are processed. This is not
compatible with the diff-mode. We need to force reload all clagd
interfaces if we detect a change.

The fix checks for a clagd change and populates the run queue with
clagd interface as needed.

Signed-off-by: Julien Fortin <jfortin@nvidia.com>