Flush DHCP client lease files on boot up
If dhclient starts and...

* we still have a valid lease file, and
* we see no responses to a series of DHCP Requests for the IP in that lease, and
* no responses to a series DHCP Discovers.

... then we try and use our old address from the last valid lease.

The address is added back to the interface, however if the router defined in the lease is not reachable through a ping check, it removes the address again. If there is no router, or the ping succeeds, the address remains.

The dhclient application does then not try and do a Discovery again until the lease is due to expire. In the event of the ping failure of the router address and the valid lease is a long one, we may not be able to access the switch for an equally long time.

This flow however only happens when the switch is unexpectedly power-cycled. In the case of a reboot, we delete the lease files on the way down so that we start fresh on next boot.

This patch here provides parity between the expected reboot and non-expected power-cycle cases by also deleting the files on boot up before we bring up mgmt class interfaces.

Signed-off-by: Julien Fortin <jfortin@nvidia.com>