Don't fork/exec/cat/grep for evaluating ifupdown2.conf
This avoids gratuitous execution of POSIX utilities in a forked shell to
do a job that python itself is perfectly capable of doing.

Note that it changes existing behavior *slightly* (but for the better),
as before this patch, even a commented out line like the following:

    # enable_persistent_debug_logging=yes

... in `ifupdown2.conf` would have led to the persistent debugging
feature being enabled, despite `ifupdown2.conf`'s content suggesting
that this is against the (afaict not formally specified) parsing rules.